requirements.txt
telethon
python-dotenv
pandas
qrcode
